rename: utils 패키지를 공통 패키지로 이동
