+++ src/main/java/com/recipe/app/common/utils/ValidationRegex.java
@@ -1,4 +1,4 @@
-package com.recipe.app.utils;
+package com.recipe.app.common.utils;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
