+++ src/main/java/com/recipe/app/src/fridge/FridgeController.java
@@ -5,7 +5,7 @@ import com.recipe.app.common.exception.BaseException;
 import com.recipe.app.common.response.BaseResponse;
 import com.recipe.app.src.fridge.models.*;
 import com.recipe.app.src.user.UserProvider;
-import com.recipe.app.utils.JwtService;
+import com.recipe.app.common.utils.JwtService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
