+++ src/main/java/com/recipe/app/src/ingredientCategory/IngredientCategoryProvider.java
@@ -2,7 +2,7 @@ package com.recipe.app.src.ingredientCategory;
 
 import com.recipe.app.common.exception.BaseException;
 import com.recipe.app.src.ingredientCategory.models.IngredientCategory;
-import com.recipe.app.utils.JwtService;
+import com.recipe.app.common.utils.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
