+++ src/main/java/com/recipe/app/src/recipeInfo/RecipeInfoController.java
@@ -7,7 +7,7 @@ import com.recipe.app.common.response.BaseResponse;
 import com.recipe.app.src.recipeInfo.models.GetRecipeBlogsRes;
 import com.recipe.app.src.recipeInfo.models.GetRecipeInfoRes;
 import com.recipe.app.src.recipeInfo.models.GetRecipeInfosRes;
-import com.recipe.app.utils.JwtService;
+import com.recipe.app.common.utils.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
