+++ src/main/java/com/recipe/app/src/recipeInfo/RecipeInfoService.java
@@ -1,6 +1,6 @@
 package com.recipe.app.src.recipeInfo;
 
-import com.recipe.app.utils.JwtService;
+import com.recipe.app.common.utils.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
