+++ src/main/java/com/recipe/app/src/recipeProcess/RecipeProcessController.java
@@ -1,6 +1,6 @@
 package com.recipe.app.src.recipeProcess;
 
-import com.recipe.app.utils.JwtService;
+import com.recipe.app.common.utils.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
