+++ src/main/java/com/recipe/app/src/scrapPublic/ScrapPublicProvider.java
@@ -7,7 +7,7 @@ import com.recipe.app.src.scrapPublic.models.ScrapPublic;
 import com.recipe.app.src.scrapPublic.models.ScrapPublicList;
 import com.recipe.app.src.user.UserProvider;
 import com.recipe.app.src.viewPublic.ViewPublicRepository;
-import com.recipe.app.utils.JwtService;
+import com.recipe.app.common.utils.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.recipe.app.common.exception.BaseException;
