+++ src/main/java/com/recipe/app/src/scrapYoutube/ScrapYoutubeProvider.java
@@ -8,7 +8,7 @@ import com.recipe.app.src.scrapYoutube.models.ScrapYoutubeList;
 import com.recipe.app.src.user.UserProvider;
 import com.recipe.app.src.user.models.User;
 import com.recipe.app.src.viewYoutube.ViewYoutubeRepository;
-import com.recipe.app.utils.JwtService;
+import com.recipe.app.common.utils.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
