+++ src/main/java/com/recipe/app/src/scrapYoutube/ScrapYoutubeService.java
@@ -4,7 +4,7 @@ import com.recipe.app.common.exception.BaseException;
 import com.recipe.app.src.scrapYoutube.models.*;
 import com.recipe.app.src.user.UserProvider;
 import com.recipe.app.src.user.models.User;
-import com.recipe.app.utils.JwtService;
+import com.recipe.app.common.utils.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
