+++ src/main/java/com/recipe/app/src/userRecipeIngredient/UserRecipeIngredientProvider.java
@@ -3,7 +3,7 @@ package com.recipe.app.src.userRecipeIngredient;
 import com.recipe.app.common.exception.BaseException;
 import com.recipe.app.src.userRecipe.models.MyRecipeIngredient;
 import com.recipe.app.src.userRecipeIngredient.models.UserRecipeIngredient;
-import com.recipe.app.utils.JwtService;
+import com.recipe.app.common.utils.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
