+++ src/main/java/com/recipe/app/src/userRecipePhoto/UserRecipePhotoController.java
@@ -1,8 +1,7 @@
 package com.recipe.app.src.userRecipePhoto;
 
-import com.recipe.app.utils.JwtService;
+import com.recipe.app.common.utils.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
